+++ tempest_fastapi_sdk/__init__.py
@@ -316,7 +316,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.72.0"
+__version__: str = "0.73.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",